点位替换完成，实现了部分地图展示的点位，但是并没有完善，比如添加图片，需要完善一下。然后完成案件点位的展示

+++ src/main/java/com/dimension/dao/CaseNodeMapper.java
@@ -2,6 +2,8 @@ package com.dimension.dao;
 
 import com.dimension.pojo.CaseNode;
 
+import java.math.BigDecimal;
+
 public interface CaseNodeMapper {
     int deleteByPrimaryKey(Integer id);
 
@@ -17,5 +19,5 @@ public interface CaseNodeMapper {
     
     CaseNode getCaseNodeByNodeId(Long nodeId);
     //统计升级点的个数,用于解决重复点问题
-    int countByBaseNodeId(Long nodeId);
+    BigDecimal countByBaseNodeId(Long nodeId);
 }
